Rename apiUrl.ts into service/routeService.ts
